src/requirements/dev.txt
pylama
pytest
pytest-cov
pytest-django
pytest-mock
beautifulsoup4
lxml
isort
